Move DiscoveredTests out of the pytest code.
